refactor zeppelinhub -> zeppelin logic into ZeppelinhubClient
